call server.js at each level
